Moved from inheritance to delegation for connection code
